+++ sketch-animated-core/src/commonMain/kotlin/com/github/panpf/sketch/decode/internal/animated_decode_utils.kt
@@ -17,79 +17,48 @@
 
 package com.github.panpf.sketch.decode.internal
 
-import com.github.panpf.sketch.util.rangeEquals
-import okio.ByteString.Companion.encodeUtf8
-import kotlin.experimental.and
-
-// https://developers.google.com/speed/webp/docs/riff_container
-private val WEBP_HEADER_RIFF = "RIFF".encodeUtf8().toByteArray()
-private val WEBP_HEADER_WEBP = "WEBP".encodeUtf8().toByteArray()
-private val WEBP_HEADER_VP8X = "VP8X".encodeUtf8().toByteArray()
-private val WEBP_HEADER_ANIM = "ANIM".encodeUtf8().toByteArray()
-
-// https://nokiatech.github.io/heif/technical.html
-private val HEIF_HEADER_FTYP = "ftyp".encodeUtf8().toByteArray()
-private val HEIF_HEADER_MSF1 = "msf1".encodeUtf8().toByteArray()
-private val HEIF_HEADER_HEVC = "hevc".encodeUtf8().toByteArray()
-private val HEIF_HEADER_HEVX = "hevx".encodeUtf8().toByteArray()
-
-// https://www.matthewflickinger.com/lab/whatsinagif/bits_and_bytes.asp
-private val GIF_HEADER_87A = "GIF87a".encodeUtf8().toByteArray()
-private val GIF_HEADER_89A = "GIF89a".encodeUtf8().toByteArray()
+import com.github.panpf.sketch.util.isAnimatedHeifFile
+import com.github.panpf.sketch.util.isAnimatedWebPFile
+import com.github.panpf.sketch.util.isGifFile
+import com.github.panpf.sketch.util.isHeifFile
+import com.github.panpf.sketch.util.isWebPFile
 
 /**
  * Return 'true' if the [ByteArray] contains a WebP image.
  *
  * @see com.github.panpf.sketch.animated.core.common.test.decode.internal.AnimatedDecodeUtilsTest.testIsWebP
  */
-fun ByteArray.isWebP(): Boolean =
-    rangeEquals(0, WEBP_HEADER_RIFF) && rangeEquals(8, WEBP_HEADER_WEBP)
+@Deprecated("Use isWebPFile instead", ReplaceWith("isWebPFile(this)"))
+fun ByteArray.isWebP(): Boolean = isWebPFile(this)
 
 /**
  * Return 'true' if the [ByteArray] contains an animated WebP image.
  *
  * @see com.github.panpf.sketch.animated.core.common.test.decode.internal.AnimatedDecodeUtilsTest.testIsAnimatedWebP
  */
-fun ByteArray.isAnimatedWebP(): Boolean = isWebP()
-        && rangeEquals(12, WEBP_HEADER_VP8X)
-        && (get(16) and 0b00000010) > 0
-        // Some webp images do not comply with standard protocols, obviously not GIFs but have GIF markup, here to do a fault tolerance
-        // The VP8X block is fixed at 9 bytes, plus the first 16 bytes, for a total of 25 bytes, so an anim block can only start at 25
-        && containsRiffAnimChunk(25)
+@Deprecated("Use isAnimatedWebPFile instead", ReplaceWith("isAnimatedWebPFile(this)"))
+fun ByteArray.isAnimatedWebP(): Boolean = isAnimatedWebPFile(this)
 
 /**
  * Return 'true' if the [ByteArray] contains an HEIF image. The [ByteArray] is not consumed.
  *
  * @see com.github.panpf.sketch.animated.core.common.test.decode.internal.AnimatedDecodeUtilsTest.testIsHeif
  */
-fun ByteArray.isHeif(): Boolean = rangeEquals(4, HEIF_HEADER_FTYP)
+@Deprecated("Use isHeifFile instead", ReplaceWith("isHeifFile(this)"))
+fun ByteArray.isHeif(): Boolean = isHeifFile(this)
 
 /**
  * Return 'true' if the [ByteArray] contains an animated HEIF image sequence.
  *
  * @see com.github.panpf.sketch.animated.core.common.test.decode.internal.AnimatedDecodeUtilsTest.testIsAnimatedHeif
  */
-fun ByteArray.isAnimatedHeif(): Boolean = isHeif()
-        && (rangeEquals(8, HEIF_HEADER_MSF1)
-        || rangeEquals(8, HEIF_HEADER_HEVC)
-        || rangeEquals(8, HEIF_HEADER_HEVX))
+@Deprecated("Use isAnimatedHeifFile instead", ReplaceWith("isAnimatedHeifFile(this)"))
+fun ByteArray.isAnimatedHeif(): Boolean = isAnimatedHeifFile(this)
 
 /**
  * Return 'true' if the [ByteArray] contains a GIF image.
  *
  * @see com.github.panpf.sketch.animated.core.common.test.decode.internal.AnimatedDecodeUtilsTest.testIsGif
  */
-fun ByteArray.isGif(): Boolean =
-    rangeEquals(0, GIF_HEADER_89A) || rangeEquals(0, GIF_HEADER_87A)
-
-/**
- * @see com.github.panpf.sketch.animated.core.common.test.decode.internal.AnimatedDecodeUtilsTest.testIsAnimatedWebP
- */
-private fun ByteArray.containsRiffAnimChunk(offset: Int = 0): Boolean {
-    (offset until size - WEBP_HEADER_ANIM.size).forEach {
-        if (rangeEquals(it, WEBP_HEADER_ANIM)) {
-            return true
-        }
-    }
-    return false
-}
+@Deprecated("Use isGifFile instead", ReplaceWith("isGifFile(this)"))
+fun ByteArray.isGif(): Boolean = isGifFile(this)