+++ sketch-animated-gif-koral/src/main/kotlin/com/github/panpf/sketch/decode/KoralGifDecoder.kt
@@ -29,7 +29,6 @@ import com.github.panpf.sketch.decode.internal.ImageFormat
 import com.github.panpf.sketch.decode.internal.calculateSampleSize
 import com.github.panpf.sketch.decode.internal.checkImageInfo
 import com.github.panpf.sketch.decode.internal.createInSampledTransformed
-import com.github.panpf.sketch.decode.internal.isGif
 import com.github.panpf.sketch.drawable.AnimatableDrawable
 import com.github.panpf.sketch.drawable.GifDrawableWrapperDrawable
 import com.github.panpf.sketch.fetch.FetchResult
@@ -44,6 +43,7 @@ import com.github.panpf.sketch.resize.isSmallerSizeMode
 import com.github.panpf.sketch.source.DataSource
 import com.github.panpf.sketch.util.Size
 import com.github.panpf.sketch.util.animatable2CompatCallbackOf
+import com.github.panpf.sketch.util.isGifFile
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.InternalCoroutinesApi
@@ -194,7 +194,7 @@ class KoralGifDecoder(
         }
 
         private fun isApplicable(fetchResult: FetchResult): Boolean {
-            return fetchResult.headerBytes.isGif()
+            return isGifFile(fetchResult.headerBytes)
         }
 
         override fun equals(other: Any?): Boolean {