+++ sketch-animated-gif/src/androidMain/kotlin/com/github/panpf/sketch/decode/ImageDecoderGifDecoder.kt
@@ -20,11 +20,11 @@ import android.os.Build
 import androidx.annotation.RequiresApi
 import com.github.panpf.sketch.ComponentRegistry
 import com.github.panpf.sketch.decode.internal.ImageDecoderAnimatedDecoder
-import com.github.panpf.sketch.decode.internal.isGif
 import com.github.panpf.sketch.fetch.FetchResult
 import com.github.panpf.sketch.request.RequestContext
 import com.github.panpf.sketch.request.disallowAnimatedImage
 import com.github.panpf.sketch.source.DataSource
+import com.github.panpf.sketch.util.isGifFile
 
 /**
  * Adds gif support by AnimatedImageDrawable
@@ -84,7 +84,7 @@ class ImageDecoderGifDecoder(
         }
 
         private fun isApplicable(fetchResult: FetchResult): Boolean {
-            return fetchResult.headerBytes.isGif()
+            return isGifFile(fetchResult.headerBytes)
         }
 
         override fun equals(other: Any?): Boolean {